GPU: apply CVE-2024-0671 patch [1/1]
PD#SWPL-183572

Problem:
The problem was that a small memory alias of a large memory allocation
could satisfy both conditions if its GPU VA happens to be aligned to
a large memory page. This would lead to the creation of a 2 MB mapping
even if the user only requested a smaller region, e.g. 4 kB.

Solution:
The driver now adds a third condition to the creation of a MMU ATE
at level 2: the total number of MMU page table entries to map must be
512, i.e. the user has requested a whole 2 MB region intentionally.
sync from https://scgit.amlogic.com/#/c/457015/

Verify:
valhall r44p1 devices

Change-Id: I8ea03a814065904f99d25cb45033e30784bf8cfb
Signed-off-by: yajing.li <yajing.li@amlogic.com>